Move message toolbar tests
